+++ src/GitHub/api.ts
@@ -388,3 +388,53 @@ export async function deleteGitHubRepository(repo: TRepo): Promise<boolean> {
 
     return Promise.reject(false);
 }
+
+/**
+ * Get starred repositories for the current user
+ *
+ * @export
+ * @async
+ * @returns {Promise<TRepo[]>}
+ */
+export async function getStarredGitHubRepositories(): Promise<TRepo[]> {
+    const octokit = new rest.Octokit({
+        auth: await credentials.getAccessToken(),
+    });
+
+    try {
+        const starredRepos = await octokit.paginate(octokit.activity.listReposStarredByAuthenticatedUser, (response) => {
+            return response.data;
+        });
+
+        return Promise.resolve(starredRepos as TRepo[]);
+    } catch (e: any) {
+        output?.appendLine(e.message, output.messageType.error);
+    }
+
+    return Promise.reject([]);
+}
+
+/**
+ * Get the current user's GitHub repositories
+ *
+ * @export
+ * @async
+ * @returns {Promise<TRepo[]>}
+ */
+export async function getAuthenticatedUserRepositories(): Promise<TRepo[]> {
+    const octokit = new rest.Octokit({
+        auth: await credentials.getAccessToken(),
+    });
+
+    try {
+        const repos = await octokit.paginate(octokit.repos.listForAuthenticatedUser, (response) => {
+            return response.data;
+        });
+
+        return Promise.resolve(repos as TRepo[]);
+    } catch (e: any) {
+        output?.appendLine(e.message, output.messageType.error);
+    }
+
+    return Promise.reject([]);
+}